requirements.txt
# librerias:
boto3
python-dotenv
langchain-aws
langchain 
langchain-mcp-adapters
mcp
asyncio